CookieTest: split `uri_matches()` specific tests to dedicated test class
